Implemented Data processor class

+++ CollisionDetectionSystem/FunctionalObjects/AudioHandler.cs
@@ -8,7 +8,7 @@ namespace CollisionDetectionSystem
 
 		//get time until intersection
 		//call threat level
-		public void OnAircraftWillIntersectInTimeEvent (float time)
+		public void OnAircraftWillIntersectInTimeEvent (double time)
 		{
 			Threat theLevel = DetermineThreatLevel (time);
 			PlayAudio (theLevel);
@@ -17,7 +17,7 @@ namespace CollisionDetectionSystem
 
 		//determines threat level and sends into playAudio if needed
 		//time is in number seconds
-		public Threat DetermineThreatLevel (float time)
+		public Threat DetermineThreatLevel (double time)
 		{
 			
 			//case for theat level determined  <30 for orange, < 60 for yellow, < 15 for red
